Adds approval proposal type
